doubles kleiner gemaakt
git-svn-id: https://svn.science.uu.nl/repos/sci.boon0104.StageHvA/IVMdraw-parent/trunk@51 a766e028-ca59-4e03-b166-cfe2b108fda4

+++ IVMdrawGWT/src/main/java/fi/ivmdrawgwt/client/Stroke.java
@@ -62,8 +62,8 @@ public class Stroke {
 		}
 		else if(parsePointsIntX!=null) {
 			for (int i = 0; i < parsePointsIntX.size(); i++) {
-				double x = ((Number)(parsePointsIntX.get(i)/10)).doubleValue();
-				double y = ((Number)(parsePointsIntY.get(i)/10)).doubleValue();
+				double x = ((Number)((double)parsePointsIntX.get(i)/10)).doubleValue();
+				double y = ((Number)((double)parsePointsIntY.get(i)/10)).doubleValue();
 				parsePoints.add(new DoublePoint(x,y));
 			}
 		}
